Refactor and split lib.rs
