+++ src/config.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use serde::{Deserialize, Serialize};
 
-use crate::{BusStopHandler, ServiceAlertHandler, SubwayStopHandler};
+use crate::{bus_stop_handler::BusStopHandler, delay::ServiceAlertHandler, subway_stop_handler::SubwayStopHandler};
 
 #[derive(Debug, Serialize, Deserialize, Default)]
 pub struct Conf {
